feat(page-title): move title experiment translation into localisation file
