+++ public/locales/en/common.json
@@ -16,7 +16,7 @@
     "settings": "Settings"
   },
   "experiences": {
-    "marquee": ["📣 Come back please 🏃‍♀️🏃 We have candy!! 🚐"],
-    "array_paged": ["⭐️ HEY YOU 🫵", "😜 YES YOU 😱", "📣 COME BACK 🏃"]
+    "marquee_variants": ["📣 Come back please 🏃‍♀️🏃 We have candy!! 🚐"],
+    "array_paged_variants": ["⭐️ HEY YOU 🫵", "😜 YES YOU 😱", "📣 COME BACK 🏃"]
   }
 }
