+++ src/features/page_title/components/PageTitleExperience.tsx
@@ -1,3 +1,5 @@
+import { useTranslation } from 'next-i18next';
+
 import { useAppSelector } from '@/redux/hooks';
 import {
   selectInteractionUnlocked,
@@ -5,7 +7,7 @@ import {
 } from '@/redux/selectors/runtime';
 import { selectPageTitle } from '@/redux/selectors/experience';
 
-import ArrayPagedTitle from './PagedTitle';
+import ArrayPagedTitle from './ArrayPagedTitle';
 import MarqueeTitle from './MarqueeTitle';
 import GlitchyTitle from './GlitchyTitle';
 
@@ -17,13 +19,30 @@ const PageTitleExperience = () => {
   const pageTitleExperience = useAppSelector(selectPageTitle);
   const isVisible = useAppSelector(selectIsDocumentVisible);
   const hasInteracted = useAppSelector(selectInteractionUnlocked);
+  const { t } = useTranslation('common');
+
+  // Using arrays in language might be stretching how i18n should be used 😰
+  // Anyways, in case of returnObjects being true we need to first convert
+  // the magic object into an array.
+  const marqueeVariants = Array.from(
+    t('experiences.marquee_variants', {
+      returnObjects: true,
+      defaultValue: [],
+    }),
+  ) as string[];
+  const arrayPagedVariants = Array.from(
+    t('experiences.array_paged_variants', {
+      returnObjects: true,
+      defaultValue: [],
+    }),
+  ) as string[];
 
   return (
     <>
-      {pageTitleExperience.inactiveMarquee && (
+      {pageTitleExperience.inactiveMarquee && marqueeVariants.length > 0 && (
         <MarqueeTitle
           enabled={hasInteracted && !isVisible}
-          text="📣 Come back please 🏃‍♀️🏃 We have candy!! 🚐"
+          text={marqueeVariants[0]}
         />
       )}
       {pageTitleExperience.randomGlitch && (
@@ -32,7 +51,7 @@ const PageTitleExperience = () => {
       {pageTitleExperience.inactiveArrayPaged && (
         <ArrayPagedTitle
           enabled={hasInteracted && !isVisible}
-          texts={['⭐️ HEY YOU 🫵', '😜 YES YOU 😱', '📣 COME BACK 🏃']}
+          texts={arrayPagedVariants}
         />
       )}
     </>
